upgrade aiq-1.1.0 to nat-1.2.1 and rename aiq to nat
