+++ pyproject.toml
@@ -19,7 +19,7 @@ dependencies = {file = ["requirements.txt"]}
 where = ["src"]
 include = ["*"]
 
-[project.entry-points.'aiq.components']
+[project.entry-points.'nat.components']
 sast_agent_workflow = "sast_agent_workflow.register"
 
 [project.scripts]
