+++ src/__init__.py
@@ -5,5 +5,5 @@ This package contains the complete SAST AI workflow system including:
 - SAST agent workflow implementation using LangGraph
 - Data models and configuration management
 - File handlers and utilities
-- Integration with AIQ toolkit
+- Integration with NAT
 """
