+++ src/sast_agent_workflow/tools/data_fetcher.py
@@ -4,10 +4,10 @@ from typing import Dict, List
 
 from pydantic import Field
 
-from aiq.builder.builder import Builder
-from aiq.builder.function_info import FunctionInfo
-from aiq.cli.register_workflow import register_function
-from aiq.data_models.function import FunctionBaseConfig
+from nat.builder.builder import Builder
+from nat.builder.function_info import FunctionInfo
+from nat.cli.register_workflow import register_function
+from nat.data_models.function import FunctionBaseConfig
 
 from dto.SASTWorkflowModels import SASTWorkflowTracker, PerIssueData
 from handlers.repo_handler_factory import repo_handler_factory
