+++ src/sast_agent_workflow/tools/pre_process.py
@@ -7,10 +7,10 @@ from common.constants import DEFAULT_FIELD_VALUE
 from common.config import Config
 from pydantic import Field
 
-from aiq.builder.builder import Builder
-from aiq.builder.function_info import FunctionInfo
-from aiq.cli.register_workflow import register_function
-from aiq.data_models.function import FunctionBaseConfig
+from nat.builder.builder import Builder
+from nat.builder.function_info import FunctionInfo
+from nat.cli.register_workflow import register_function
+from nat.data_models.function import FunctionBaseConfig
 
 from dto.LLMResponse import AnalysisResponse, CVEValidationStatus, FinalStatus
 from dto.SASTWorkflowModels import PerIssueData, SASTWorkflowTracker
