+++ tests/nat_tests/test_filter.py
@@ -13,8 +13,8 @@ from dto.Issue import Issue
 from dto.LLMResponse import CVEValidationStatus
 from dto.ResponseStructures import KnownFalsePositive
 from common.config import Config
-from aiq.builder.builder import Builder
-from tests.aiq_tests.test_utils import TestUtils
+from nat.builder.builder import Builder
+from tests.nat_tests.test_utils import TestUtils
 
 
 class TestFilterCore(unittest.IsolatedAsyncioTestCase):
