+++ tests/nat_tests/test_write_results.py
@@ -10,8 +10,8 @@ from dto.SASTWorkflowModels import SASTWorkflowTracker
 from dto.Issue import Issue
 from dto.LLMResponse import AnalysisResponse
 from common.config import Config
-from aiq.builder.builder import Builder
-from tests.aiq_tests.test_utils import TestUtils
+from nat.builder.builder import Builder
+from tests.nat_tests.test_utils import TestUtils
 
 
 class TestWriteResultsCore(unittest.IsolatedAsyncioTestCase):
